requirements.txt
uvicorn==0.23.2
fastapi==0.104.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
alembic==1.12.1
pydantic-settings==2.0.3
bcrypt==4.0.1
python-jose==3.3.0
httpx==0.25.1
pillow==10.1.0
aiofiles==23.2.1
cloudinary==1.36.0
python-multipart==0.0.6
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.23.2
Jinja2==3.1.2
aiofiles==23.2.1
openai==1.6.1
celery==5.3.6